refactoring mocks
